Rename and move refactors
-  agGridColumnSizing.ts -> columnSizing.ts
-  the other agGrid* files renamed to just ag*
   -  only small wrappers and utilities on top of ag-grid are now called ag*
-  moved one misplaced test

JIRA: FET-608
